Got query functioning from interface Had to play with params so that it would return the correct thing Must enter query= to pass in a search item in the navbar
