+++ app/views/searches/index.html.erb
@@ -1,3 +1,1 @@
-<% @results.each do |result| %>
-  <%= puts result %>
-  <% end %>
+<%= @results.first %>
